backend/requirements.txt
Flask
flask-cors
openai==0.28
PyMuPDF
spacy
python-dotenv==1.0.0
